requirements.txt
flask==2.3.2
flask-cors==3.0.10
nltk==3.8.1
numpy==1.24.3
pandas==2.0.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.2.2
torch==2.0.1
cohere==4.3.1
gunicorn==20.1.0